Rewrote the procedure on Automated Tests
The former procedure described test-driven development; but our team doesn't strictly adhere to TDD. If the team adopts that practice, than _that_  should be added to our Desktop Procedures; but in the mean time, this procedure should reflect our current best-practice.

I believe we currently recognize several things:

 1. That some of our code is _under_-tested
 2. That some of our tests can be culled (or improved)
 3. That we don't aspire to 100% code coverage; but consider testing to be more of a cost-benefit tradeoff.

My rewrite of this procedure seeks to articulate our understanding that writing tests is a matter of balancing tradeoffs versus a one-size-fits-all prescription. It also seeks to articulate the specific costs and benefits of the tradeoff so that, as we write, rewrite, or cull tests, we can evaluate them in light of a clear rubric.