Updated and Edited the procedure for creating commits
The old procedure discussed `ep merge` which probably shouldn't be in this document for two reasons:

 1. We no longer use that tool
 2. It belongs more to the Deploying Changes "chapter" than the Developing Features one — this procedure is applicable to any developer; merging Pull Requests is applicable particularly to developers acting in the role of Project Maintainer.

I also edited the document down to a smaller word count in the effort to keep the repo succinct.

In editing, I chose to focus on the higher-level content that succinctly encodes our rules and conventions — content that's perennially of interest to any developer in EP — as opposed to content that you might learn once and have no need of re-referencing (like when you'll need to use `-f` to force-push or that you might want to `ep unpair` after pairing).